+++ babel.config.js
@@ -20,6 +20,7 @@ module.exports = function(api) {
                   '@edtr-io/plugin-anchor': '@edtr-io/plugin-anchor/src',
                   '@edtr-io/plugin-blockquote':
                     '@edtr-io/plugin-blockquote/src',
+                  '@edtr-io/plugin-exercise': '@edtr-io/plugin-exercise/src',
                   '@edtr-io/plugin-equations': '@edtr-io/plugin-equations/src',
                   '@edtr-io/plugin-files': '@edtr-io/plugin-files/src',
                   '@edtr-io/plugin-geogebra': '@edtr-io/plugin-geogebra/src',