+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased](https://github.com/edtr-io/edtr-io/compare/v0.6.0..HEAD)
+
+### Added
+
+- **plugin-exercise**. Plugin for Exercises with text and solution.
+
 ## [0.6.0](https://github.com/edtr-io/edtr-io/compare/v0.5.0..v0.6.0)
 
 ### Breaking Changes